fix: respect worker package file-count baselines
